LayoutTests/inspector/styles/styles-add-new-rule-expected.txt

Tests that adding a new rule works after switching nodes.

Text
Revision added: inspector-stylesheet
After adding new rule (inspected):
[expanded] 
element.style  { ()
font-size: 12px;

======== Matched CSS Rules ========
[expanded] 
foo, [$#inspected$], bar  { (inspector-stylesheet:1)
color: maroon;

[expanded] 
div  { (user agent stylesheet)
display: block;


After adding new rule (other):
[expanded] 
element.style  { ()
color: red;

======== Matched CSS Rules ========
[expanded] 
#other  { (inspector-stylesheet:1)

[expanded] 
div  { (user agent stylesheet)
display: block;
